Labour Market Overview Tables

A dataset containing selected content from the Labour Market Overview Tables publication (UK, December 2020), published by the Office for National Statistics as a demonstration of good practice for releasing statistics in spreadsheets (June 2021).

lfs_tables

Format

A data frame with 6 rows and 7 columns:

tab_title

Character. Text to appear on the sheet's tab.

sheet_type

Character. Either "meta" or "tables".

sheet_title

Character. Title that will appear in the top-left of each sheet

source

Character. The origin of the data.

subtable_num

Character. A letter. Identifier for tables when there's more than one per sheet.

subtable_title

Character. A name to place above the table if there's more than one table per sheet.

table_name

Character. A name applied to the marked-up table.

table

List containing dataframes/tibbles. Rectangular table containing the data.

Source

https://gss.civilservice.gov.uk/policy-store/releasing-statistics-in-spreadsheets/