Create An a11ytable-class Object

Create an object containing all the input data and supporting information for your publication, which will be used to populate a workbook.

new_a11ytable(
  tab_titles,
  sheet_types = c("cover", "contents", "notes", "tables"),
  sheet_titles,
  sources = NULL,
  subtable_nums = NULL,
  subtable_titles = NULL,
  table_names,
  tables
)

Arguments

tab_titles

Required character vector, one value per sheet. The text that will appear on the tab interface of the when opened in spreadsheet software. Provide table sheets as numbers.

sheet_types

Required character vector, one value per sheet. Meta sheets (i.e. those that don't contain publication tables) should be 'contents', 'cover' or 'notes'; publication tables are type 'tables'.

sheet_titles

Required character vector, one value per sheet. The main title of the sheet. Will appear in large text in cell A1 (top-left) corner of each sheet.

sources

Optional character vector, one value per sheet (NA is valid). The origin of the data for a given sheet.

subtable_nums

Not currently implemented. Character vector. The table numbers for when a sheet contains more than one table (not recommended).

subtable_titles

Not currently implemented. Character vector. The table titles for when a sheet contains more than one table (not recommended).

table_names

Required character vector, one value per sheet. A name to give the 'named range' of cells that compose the table in the final spreadsheet output. Makes sheet navigation easier.

tables

Required list of data.frames, one per sheet. See details.

Value

An a11ytables-class object.

Details

Formats for data.frames in the 'tables' argument, depending on the sheet type:

cover: one column with a pair of rows for each point of information (a section with contact details should have a row for the title of that section, like 'Contact details' and one row containing the information; use line breaks to separate information onto different lines, like a telephone number and email address in this example)

contents: one row per sheet, five columns suggested ('Worksheet number', 'Worksheet title', 'Data of first publication', 'Date of next publication', 'Source')

notes: one row per note, two columns suggested ('Note number', 'Note text')

tables: a tidy, rectangular data.frame containing the data to be published

Examples

if (FALSE) {
x <- new_a11ytable(
    tab_titles      = lfs_tables$tab_title,
    sheet_types     = lfs_tables$sheet_type,
    sheet_titles    = lfs_tables$sheet_title,
    sources         = lfs_tables$source,
    subtable_nums   = lfs_tables$subtable_num,
    subtable_titles = lfs_tables$subtable_title,
    table_names     = lfs_tables$table_name,
)
is_a11ytable(x)
}